Modified all commands with abstract class Command.java
Modified all system buses with abstract class SystemBus.java

+++ src/cpu_sim/bus/ControlBus.java
@@ -10,12 +10,12 @@ public class ControlBus extends SystemBus{
     }
 
     @Override
-    public byte getBus16bit() {
+    public short getBus16bit() {
         return super.getBus16bit();
     }
 
     @Override
-    public void setBus16bit(byte bus16bit) {
+    public void setBus16bit(short bus16bit) {
         super.setBus16bit(bus16bit);
     }
 }
